feat: finalize Event Resource Guide content model and schema
- Centralized content type definitions in `src/lib/types/content.ts`.
- Documented canonical frontmatter schema in `event-resource-guide-format.md`.
- Added `whyAttending` support for Hero sections and JSON-LD schema.
- Support nested YAML for `theme` and `gear` structures with flat fallbacks.
- Created `content/events/sample-event.md` as a comprehensive fixture.
- Updated `tests/test-constants.ts` to ignore external cross-origin noise in CI.
