+++ content/events/jack-and-jill-orama.md
@@ -9,12 +9,6 @@ excerpt: "The ultimate West Coast Swing party and competition weekend in Souther
 location: "Hyatt Regency Orange County"
 city: "Garden Grove, CA"
 schedule: "June 4 - 7, 2026"
-region: "Pacific Northwest & West Coast"
-resourceGuide: "/events/jack-and-jill-orama"
-guideStatus:
-  - "Live Guide"
-  - "Theme Gear"
-  - "Reminder Alerts"
 url: "https://jackandjillorama.com"
 heroImage: ""
 description: "Organized by Ben Morris, Jack & Jill O'Rama is one of the most popular events on the circuit. It is famous for its creative competition formats, high-energy social dancing, and prime location near Disneyland."
