+++ content/events/soswing.md
@@ -9,11 +9,6 @@ excerpt: "A charming WCS experience in beautiful Ashland, OR."
 location: "Ashland Hills Hotel & Suites"
 city: "Ashland, OR"
 schedule: "May 14 - 17, 2026"
-region: "Pacific Northwest"
-guideStatus:
-  - "Live Guide"
-  - "Theme Gear"
-  - "Reminder Alerts"
 url: "https://soswing.com"
 heroImage: ""
 description: "SOswing offers a unique, community-focused atmosphere in beautiful Ashland. Known for its 'Westie' hospitality, the event features WSDC Jack & Jill competitions, leveled workshops with top-tier pros, and late-night social dancing."
